requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-google==10.3.0
soda-core-bigquery==3.0.45
astro-run-dag # This package is needed for the astro run command. It will be removed before a deploy